+++ packages/wallets/keepkey/src/chains/utxo.ts
@@ -13,13 +13,14 @@ import {
 } from '@coinmasters/toolbox-utxo';
 import type { UTXOChain } from '@coinmasters/types';
 import { Chain, DerivationPath, FeeOption } from '@coinmasters/types';
-import { xpubConvert } from '@pioneer-platform/pioneer-coins';
-import { toCashAddress } from 'bchaddrjs';
-import type { Psbt } from 'bitcoinjs-lib';
 import {
   ChainToNetworkId,
   // @ts-ignore
 } from '@pioneer-platform/pioneer-caip';
+import { xpubConvert } from '@pioneer-platform/pioneer-coins';
+import { toCashAddress } from 'bchaddrjs';
+import type { Psbt } from 'bitcoinjs-lib';
+
 import { bip32ToAddressNList, ChainToKeepKeyName } from '../helpers/coins.ts';
 const TAG = ' | kk-utxo | ';
 type Params = {
@@ -86,25 +87,40 @@ export const utxoWalletMethods = async ({
 
   const scriptType = segwit ? 'p2sh' : 'p2pkh';
   if (!ChainToKeepKeyName[chain]) throw Error('ChainToKeepKeyName: unknown chain: ' + chain);
-  const addressInfo = {
-    coin: ChainToKeepKeyName[chain],
-    script_type: scriptType,
-    address_n: bip32ToAddressNList(DerivationPath[chain]),
+  // const addressInfo = {
+  //   coin: ChainToKeepKeyName[chain],
+  //   script_type: scriptType,
+  //   address_n: bip32ToAddressNList(DerivationPath[chain]),
+  // };
+  // //console.log('addressInfo: ', addressInfo);
+  // const { address: walletAddress } = await sdk.address.utxoGetAddress(addressInfo);
+
+  const getAddress = async (customAddressInfo?: any) => {
+    let tag = TAG + ' | getAddress | ';
+    const defaultAddressInfo = {
+      coin: ChainToKeepKeyName[chain],
+      script_type: scriptType,
+      address_n: bip32ToAddressNList(DerivationPath[chain]),
+    };
+    const addressInfo = customAddressInfo || defaultAddressInfo;
+    console.log(tag, 'addressInfo: ', addressInfo);
+    const { address: walletAddress } = await sdk.address.utxoGetAddress(addressInfo);
+    return walletAddress as string;
   };
-  //console.log('addressInfo: ', addressInfo);
-  const { address: walletAddress } = await sdk.address.utxoGetAddress(addressInfo);
 
   // @ts-ignore
   const _getPubkeys = async (paths: any) => {
+    let tag = TAG + ' | _getPubkeys | ';
     try {
+      console.log(tag, 'Checkpoint 1');
       if (!paths || !paths.length) return [];
       console.time('getPubkeys Duration' + chain); // Starts the timer
       let pubkeys = await Promise.all(
         paths.map(async (path: any) => {
-          console.log('path: ', path);
+          console.log(tag, 'path: ', path);
           if (!path.addressNList) throw new Error('addressNList not found in path: FATAL');
           if (path.type === 'address') return;
-          if (path.script_type === 'p2wpkh' && path.coin !== 'Bitcoin') return;
+          // if (path.script_type === 'p2wpkh' && path.coin !== 'Bitcoin') return;
           // Marked as async to use await inside
           // Create the path query for public key retrieval
           const pathQuery = {
@@ -116,9 +132,9 @@ export const utxoWalletMethods = async ({
           };
 
           // Get the public key
-          console.log('pathQuery: ', pathQuery);
+          console.log(tag, 'pathQuery: ', pathQuery);
           const pubkeyResponse = await sdk.system.info.getPublicKey(pathQuery);
-          console.log('pubkeyResponse: ', pubkeyResponse);
+          console.log(tag, 'pubkeyResponse: ', pubkeyResponse);
           if (!pubkeyResponse) throw new Error('Failed to get response from keepkey: FATAL');
           if (!pubkeyResponse.xpub)
             throw new Error('Failed to get correct response from keepkey: FATAL');
@@ -131,11 +147,14 @@ export const utxoWalletMethods = async ({
           };
 
           // Get the master address
+          console.log(tag, 'addressInfo: ', addressInfo);
           const addressMaster = await sdk.address.utxoGetAddress(addressInfo);
 
           if (path.script_type === 'p2wpkh') {
             //convert xpub to zpub
             pubkeyResponse.xpub = xpubConvert(pubkeyResponse.xpub, 'zpub');
+            if (!pubkeyResponse.xpub) throw Error('Failed to format xpub to zpub: FATAL');
+            console.log('converted from zpub: pubkeyResponse.xpub: ', pubkeyResponse.xpub);
           }
 
           // Combine the original path object with the xpub and master from the responses
@@ -143,6 +162,7 @@ export const utxoWalletMethods = async ({
             ...path, // Contains all fields from the original path
             xpub: pubkeyResponse.xpub, // Adds the xpub field from the response
             master: addressMaster.address, // Adds the master address from the response
+            address: addressMaster.address, // Adds the master address from the response
           };
         }),
       );
@@ -161,6 +181,8 @@ export const utxoWalletMethods = async ({
 
   const signTransaction = async (psbt: Psbt, inputs: KeepKeyInputObject[], memo: string = '') => {
     //console.log('psbt.txOutputs: ', psbt.txOutputs);
+    let walletAddress = await getAddress();
+
     const outputs = psbt.txOutputs
       .map((output) => {
         const { value, address, change } = output as psbtTxOutput;
@@ -240,8 +262,9 @@ export const utxoWalletMethods = async ({
 
       // Filter the pubkeys by networkId
       let filterPubkeys = pubkeys.filter((pubkey: any) =>
-        pubkey.networks.some((network: any) => network === ChainToNetworkId[chain])
+        pubkey.networks.some((network: any) => network === ChainToNetworkId[chain]),
       );
+      console.log(tag, 'filterPubkeys: ', filterPubkeys);
       const { psbt, inputs: rawInputs } = await toolbox.buildTx({
         ...rest,
         pubkeys: filterPubkeys,
@@ -316,7 +339,7 @@ export const utxoWalletMethods = async ({
     ...toolbox,
     getPubkeys,
     // getInputsForPubkey,
-    getAddress: () => walletAddress as string,
+    getAddress,
     signTransaction,
     transfer,
   };