+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/toolbox-cosmos
 
+## 11.0.31
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.31
+  - @coinmasters/api@3.7.31
+
 ## 11.0.30
 
 ### Patch Changes