+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/wallet-keplr
 
+## 11.0.31
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.31
+  - @coinmasters/types@4.7.31
+
 ## 11.0.30
 
 ### Patch Changes